+++ e2e-test/cypress/integration/subproject_create_spec.js
@@ -1,6 +1,7 @@
 describe("Subproject creation", function() {
+  let projectId;
+
   it("Root cannot add a subproject", function() {
-    let projectId;
     const organization = "ACME Corp";
     const projectProjectedBudget = {
       organization,
@@ -21,5 +22,30 @@ describe("Subproject creation", function() {
       .then(() => cy.visit(`/projects/${projectId}`))
       .then(() => cy.get("[data-test=subproject-create-button]").should("be.visible"))
       .then(() => cy.get("[data-test=subproject-create-button]").should("be.disabled"));
+  });
+
+  it("Check warnings that permissions are not assigned", function() {
+    cy.login();
+    cy.visit(`/projects/${projectId}`);
+
+    //Create a subproject
+    cy.get("[data-test=subproject-create-button]").click();
+    cy.get("[data-test=nameinput] input").type("Test");
+    cy.get("[data-test=dropdown-sp-dialog-currencies-click]")
+      .click()
+      .then(() => cy.get("[data-value=EUR]").click());
+    cy.get("[data-test=submit]").click();
+
+    //Check snackbar warning visible
+    cy.get("[data-test=client-snackbar]")
+      .should("be.visible")
+      .should("contain", "permissions");
+
+    //Check warning badge
+    cy.get("[data-test=warning-badge]").should("be.visible");
+    cy.get("[data-test=spp-button-0]").click();
+    cy.get("[data-test=warning-badge]").should("not.be.checked");
+    cy.get("[data-test=permission-submit]").click();
+    cy.get("[data-test=warning-badge]").should("not.be.checked");
   });
 });