+++ frontend/src/pages/SubProjects/reducer.js
@@ -56,6 +56,7 @@ const defaultState = fromJS({
     currency: "",
     projectedBudgets: []
   },
+  idsPermissionsUnassigned: [],
   creationDialogShown: false,
   editDialogShown: false,
   showHistory: false,
@@ -113,7 +114,8 @@ export default function detailviewReducer(state = defaultState, action) {
         temporaryPermissions: defaultState.get("temporaryPermissions"),
         idForPermissions: action.id,
         displayNameForPermissions: action.displayName,
-        showSubProjectPermissions: true
+        showSubProjectPermissions: true,
+        idsPermissionsUnassigned: state.get("idsPermissionsUnassigned").filter(id => id !== action.id)
       });
     case SHOW_SUBPROJECT_ADDITIONAL_DATA:
       return state.merge({
@@ -162,7 +164,8 @@ export default function detailviewReducer(state = defaultState, action) {
     case CREATE_SUBPROJECT_SUCCESS:
       return state.merge({
         highlightingRegex: defaultState.get("highlightingRegex"),
-        subprojectToAdd: defaultState.get("subprojectToAdd")
+        subprojectToAdd: defaultState.get("subprojectToAdd"),
+        idsPermissionsUnassigned: state.get("idsPermissionsUnassigned").concat(fromJS(action.subprojectId))
       });
     case SHOW_PROJECT_ASSIGNEES:
       return state.set("showProjectAssignees", true);