+++ frontend/src/pages/Workflows/WorkflowItem.js
@@ -18,6 +18,7 @@ import _isEmpty from "lodash/isEmpty";
 import React from "react";
 import { SortableElement } from "react-sortable-hoc";
 import ActionButton from "../Common/ActionButton";
+import StyledBadge from "../Common/StyledBadge";
 
 import { amountTypes, fromAmountString, toAmountString } from "../../helper.js";
 import strings from "../../localizeStrings";
@@ -79,7 +80,7 @@ const styles = {
     bottom: "34px"
   },
 
-  infoButton: {
+  buttonStyle: {
     minWidth: "30px",
     marginLeft: "5px"
   },
@@ -215,6 +216,7 @@ function isWorkflowItemSelectable(redacted, sortenabled, allowedIntents) {
   // a user must have assign permissions or all three permission handling permissions
   return allowedIntents.includes("workflowitem.assign") || intents.length === 3 ? true : false;
 }
+
 const editWorkflow = (
   { id, displayName, amount, exchangeRate, amountType, currency, description, status, documents },
   props
@@ -240,7 +242,7 @@ const getInfoButton = ({ openWorkflowDetails }, status, workflowSortEnabled, wor
     <IconButton
       className="workflowitem-info-button"
       disabled={workflowSortEnabled}
-      style={{ ...getButtonStyle(workflowSortEnabled, status), ...styles.infoButton }}
+      style={{ ...getButtonStyle(workflowSortEnabled, status), ...styles.buttonStyle }}
       onClick={() => openWorkflowDetails(workflow.id)}
       data-test={`workflowitem-info-button-${workflow.id}`}
     >
@@ -320,13 +322,17 @@ const renderActionButtons = (
   workflowSortEnabled,
   status,
   showAdditionalData,
-  additionalData
+  additionalData,
+  idsPermissionsUnassigned,
+  id
 ) => {
   const additionalDataDisabled = _isEmpty(additionalData) || workflowSortEnabled;
   const editDisabled = !canEditWorkflow || workflowSortEnabled;
   const permissionsDisabled = !canListWorkflowPermissions || workflowSortEnabled;
-
+  const workflowitemAssigneeChanged = idsPermissionsUnassigned.find(el => el === id) === undefined;
+  const isBadgeHidden = workflowitemAssigneeChanged || workflowSortEnabled || permissionsDisabled ? true : false;
   const closeDisabled = !canCloseWorkflow || workflowSortEnabled;
+  const permissionsTitle = isBadgeHidden ? strings.common.show_permissions : strings.confirmation.assign_permissions;
   return (
     <div style={styles.actionCell}>
       <div style={styles.actions}>
@@ -350,16 +356,24 @@ const renderActionButtons = (
           data-test="edit-workflowitem"
           iconButtonStyle={getButtonStyle(workflowSortEnabled, status)}
         />
-        <ActionButton
-          notVisible={workflowSortEnabled || permissionsDisabled}
-          onClick={permissionsDisabled ? undefined : showPerm}
-          icon={<PermissionIcon />}
-          title={permissionsDisabled ? "" : strings.common.show_permissions}
-          workflowSortEnabled={workflowSortEnabled}
-          status={status}
-          data-test="show-workflowitem-permissions"
-          iconButtonStyle={getButtonStyle(workflowSortEnabled, status)}
-        />
+        <StyledBadge
+          color="secondary"
+          variant="dot"
+          invisible={isBadgeHidden}
+          data-test={"warning-badge"}
+          style={{ ...styles.buttonStyle }}
+        >
+          <ActionButton
+            notVisible={workflowSortEnabled || permissionsDisabled}
+            onClick={permissionsDisabled ? undefined : showPerm}
+            icon={<PermissionIcon />}
+            title={permissionsDisabled ? "" : permissionsTitle}
+            workflowSortEnabled={workflowSortEnabled}
+            status={status}
+            data-test="show-workflowitem-permissions"
+            iconButtonStyle={getButtonStyle(workflowSortEnabled, status)}
+          />
+        </StyledBadge>
         <ActionButton
           notVisible={workflowSortEnabled || status === "closed" || closeDisabled}
           onClick={closeDisabled ? undefined : close}
@@ -375,7 +389,17 @@ const renderActionButtons = (
 };
 
 export const WorkflowItem = SortableElement(
-  ({ workflow, mapIndex, index, currentWorkflowSelectable, workflowSortEnabled, parentProject, users, ...props }) => {
+  ({
+    workflow,
+    mapIndex,
+    index,
+    currentWorkflowSelectable,
+    workflowSortEnabled,
+    parentProject,
+    users,
+    idsPermissionsUnassigned,
+    ...props
+  }) => {
     const { storeWorkflowItemsSelected, selectedWorkflowItems, currency: targetCurrency } = props;
     const {
       id,
@@ -396,7 +420,6 @@ export const WorkflowItem = SortableElement(
       : {
           opacity: 0.3
         };
-
     const showEdit = canUpdateWorkflowItem(allowedIntents) && status !== "closed";
     const showClose = canCloseWorkflowItem(allowedIntents) && workflowSelectable && status !== "closed";
     const infoButton = getInfoButton(props, status, workflowSortEnabled, workflow.data);
@@ -459,7 +482,9 @@ export const WorkflowItem = SortableElement(
               workflowSortEnabled,
               status,
               () => props.showWorkflowitemAdditionalData(id),
-              additionalData
+              additionalData,
+              idsPermissionsUnassigned,
+              id
             )}
           </div>
         </Card>
@@ -484,7 +509,7 @@ export const RedactedWorkflowItem = SortableElement(
         <Card elevation={workflowSelectable ? 1 : 0} key={mapIndex} style={styles.card}>
           <div style={{ ...tableStyle, ...styles.workflowContent }}>
             <div style={{ flex: 1 }}>
-              <IconButton style={styles.infoButton}>
+              <IconButton style={styles.buttonStyle}>
                 <HiddenIcon />
               </IconButton>
             </div>